1-59 All Fee Review Comments First Cut

+++ app/assets/javascripts/angular_modules/controllers/fee_types.js
@@ -10,10 +10,12 @@
             $('#createModal').modal('show')
         };
         
-        feeTypeService.index()
-            .then(function(responce){
-                $scope.fee_types = responce.data
-            });
+        $scope.index = function(){
+            feeTypeService.index()
+                .then(function(responce){
+                    $scope.fee_types = responce.data
+                });
+        }
         
 
         $scope.addMorefeeTypes = function(){
@@ -24,7 +26,7 @@
         }
         
         $scope.create = function(){
-            feeTypeService.createBulkFeeTypes($("form#createFeeTypes").serialize())
+            feeTypeService.createBulkFeeTypes($scope.new_fee_types)
                 .then(function(responce){
                     angular.forEach(responce.data, function(value) {
                         this.push(value);
@@ -34,23 +36,32 @@
         }
 
         $scope.edit = function(fee_type){
-            $scope.fee_type = fee_type
+            $scope.fee_type = angular.copy(fee_type)
             $("#editModal").modal("show")
         }
         
 
         $scope.update = function(fee_type){
-            feeTypeService.updateFeetype(fee_type)
-                .then(function(responce){
-                    $('#editModal').modal('hide')
-                });
+            $scope.error = null
+            if(typeof fee_type.fee_type != 'undefined' && fee_type.fee_type.length > 0){
+                feeTypeService.updateFeetype(fee_type)
+                    .then(function(responce){
+                        $scope.index();
+                        $('#editModal').modal('hide')
+                    });
+            }else{
+                $scope.error = "Fee Type Cant be blank"
+            }
+
         };
 
         $scope.destroy = function(fee_type){
-            feeTypeService.destroy(fee_type)
-                .then(function(responce){
-                    $scope.fee_types.splice($scope.fee_types.indexOf(fee_type), 1)
-                });
+            if(confirm("Are You Sure..??")){
+                feeTypeService.destroy(fee_type)
+                    .then(function(responce){
+                        $scope.fee_types.splice($scope.fee_types.indexOf(fee_type), 1)
+                    });
+            }
         };
     }]);
